src/requirements.txt
functions-framework==3.*
requests
pandas
pandas-gbq
google-cloud-bigquery
db-dtypes
python-dotenv==1.0.1